+++ graph-samples/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.2 (2026-09-26)
+
+### 🧱 Updated Dependencies
+
+- Updated graph-format to 1.0.5
+
 ## 0.1.1 (2026-09-24)
 
 ### 🚀 Features