+++ remote-logger/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.3.7 (2026-09-26)
+
+This was a version bump only for @graphty/remote-logger to align it with other projects, there were no code changes.
+
 ## 1.3.6 (2026-09-24)
 
 This was a version bump only for @graphty/remote-logger to align it with other projects, there were no code changes.